+++ src/main/ts/TinyMCE.ts
@@ -1,11 +1,3 @@
-/**
- * Copyright (c) 2017-present, Ephox, Inc.
- *
- * This source code is licensed under the Apache 2 license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
 import { TinyMCE as TinyMCEGlobal } from 'tinymce';
 
 const getTinymce = (view: Window): TinyMCEGlobal | null => {